@@ -5,85 +5,49 @@ import { take } from "rxjs/operators/take";
 import { mergeMap } from "rxjs/operators/mergeMap";
 import { createEpicMiddleware } from "redux-observable";
 
-import { createObservableStore } from "./store";
+import { createObservableStore, getActiveAccount$ } from "./store";
 import { createWeb3Instance } from "./utils";
 import {
   getNetwork,
   getBalance,
   getLatestBlock,
-  getDefaultAccount,
   getBlock,
-  getEvents
+  getEvents,
+  watchEvents
 } from "@eth-proxy/rx-web3";
 import {
   createSetNetwork,
   getDetectedNetwork$,
-  getContractFromNameOrAddress,
-  getActiveAccount,
-  getUniqEvents$
+  getContractFromRef
 } from "./store";
-import {
-  registerContract,
-  RegisterContractOptions
-} from "./modules/register-contract";
-import { exec } from "./modules/exec";
+import { registerContract } from "./modules/register-contract";
 import { rootEpic } from "./store/epics";
-import { call } from "./modules/call";
 import {
-  TransactionWithHash,
-  TruffleJson,
-  ConfirmedTransaction,
-  FailedTransaction,
-  ContractInfo,
   EthProxyOptions,
-  QueryModel
+  ContractRef,
+  EthProxy
 } from "./model";
-import { first, mergeMapTo } from "rxjs/operators";
-import "rxjs/add/Observable/timer";
+import { first } from "rxjs/operators";
 import { query } from "./modules/query";
-import { read, DataReader, ReadStrategy } from "./modules/read";
-import * as Web3 from 'web3';
-
-export class EthProxy {
-  registerContract: (abi, options: RegisterContractOptions) => void;
-  exec: (
-    nameOrAddress: string,
-    method: string,
-    args?: any,
-    tx_params?: any
-  ) => Observable<any>;
-  call: (
-    nameOrAddress: string,
-    method: string,
-    args?: any,
-    tx_params?: any
-  ) => Promise<any>;
-  // rxweb3
-  getBalance: (account: string) => any;
-  getLatestBlock: () => any;
-  getBlock: (args) => any;
-
-  //
-  provider$: Observable<Web3.Provider>;
-  network$: Observable<string>;
-  defaultAccount$: Observable<string | undefined>;
-
-  getContractInfo: (nameOrAddress: string) => Observable<ContractInfo>;
-  query: (queryModel: QueryModel) => Observable<any>;
-
-  events$: Observable<any[]>;
-  read: <T>(readDef: DataReader<T>, strategy?: ReadStrategy) => Observable<T>;
-}
+import * as Web3 from "web3";
+import { identity } from "ramda";
+import {
+  process,
+  createCallAdapter,
+  createExecAdapter
+} from "./modules/contract";
 
 const defaultOptions = {
   pollInterval: 1000,
-  eventReader: getEvents
+  eventReader: getEvents,
+  interceptors: {}
 };
 
-export function createProxy(
+export function createProxy<T>(
   provider$: Observable<any>,
-  options: EthProxyOptions = defaultOptions
-): EthProxy {
+  userOptions: Partial<EthProxyOptions>
+): EthProxy<T> {
+  const options = { ...defaultOptions, ...userOptions };
   const replayProvider$ = provider$.pipe(shareReplay(1), take(1));
   const web3Proxy$ = replayProvider$.pipe(map(createWeb3Instance));
 
@@ -91,42 +55,55 @@ export function createProxy(
     .pipe(mergeMap(getNetwork), map(createSetNetwork))
     .subscribe(action => store.dispatch(action));
 
+  const getEvents = (filter: Web3.FilterObject) =>
+    web3Proxy$.let(mergeMap(web3 => options.eventReader(web3, filter)));
+
+  const appWatchEvents = (filter: Web3.FilterObject) =>
+    web3Proxy$.let(mergeMap(web3 => watchEvents(web3, filter)));
+
   const epicMiddleware = createEpicMiddleware(rootEpic, {
     dependencies: {
       web3Proxy$,
-      options
+      options,
+      getEvents,
+      watchEvents: appWatchEvents
     }
   });
-  const store = createObservableStore(epicMiddleware);
+
+  const store = createObservableStore(epicMiddleware, options.store);
+
+  const interceptors = {
+    call: createCallAdapter({
+      userInterceptor: (options.interceptors as any).call || identity
+    }),
+    exec: createExecAdapter({
+      store,
+      userInterceptor:
+        (options.interceptors as any).transaction || identity
+    }),
+    preQuery: (options.interceptors as any).preQuery || identity,
+    postQuery: (options.interceptors as any).postQuery || identity
+  };
 
   return {
     provider$: replayProvider$,
     registerContract: registerContract(store),
-    
-    exec: exec(store, web3Proxy$),
-    call: call(store, web3Proxy$),
-    query: query(store, web3Proxy$, options.eventReader),
-    read: <T>(readDef: DataReader<T>, strategy?: ReadStrategy) => web3Proxy$.pipe(
-      mergeMap(web3 => read(store)(web3)(readDef, strategy))
-    ),
+
+    send: process(store, web3Proxy$, interceptors) as any,
+    query: query(store, interceptors),
 
     network$: store.let(getDetectedNetwork$),
-    defaultAccount$: store.select(getActiveAccount),
-    
-    // switch to readers
+    defaultAccount$: store.let(getActiveAccount$),
+
     getBalance: account => web3Proxy$.pipe(mergeMap(getBalance(account))),
     getLatestBlock: () => web3Proxy$.pipe(mergeMap(getLatestBlock)),
     getBlock: arg => web3Proxy$.pipe(mergeMap(getBlock(arg))),
-    getContractInfo: (nameOrAddress: string) =>
-      store
-        .select(getContractFromNameOrAddress(nameOrAddress))
-        .pipe(first(x => !!x)),
-    
-    // To be removed
-    events$: store.let(getUniqEvents$),
-
+    getContractInfo: (ref: ContractRef) =>
+      store.select(getContractFromRef(ref)).pipe(first(x => !!x))
   };
 }
 
 export * from "./model";
 export * from "./utils";
+export { ethProxyIntegrationReducer, State as EthProxyState } from "./store";
+export { on, once } from './modules/contract'
