@@ -1,10 +1,10 @@
-import { Observable } from "rxjs/Observable";
-import "rxjs/add/Observable/race";
-import "rxjs/add/Observable/of";
-import "rxjs/add/Observable/fromEvent";
 import * as Web3 from "web3";
 import { createProxy, EthProxyOptions, EthProxy } from "@eth-proxy/client";
-import { single, map, filter } from "rxjs/operators";
+import { Observable } from "rxjs/Observable";
+import { map, filter } from "rxjs/operators";
+import { race } from "rxjs/observable/race";
+import { of } from "rxjs/observable/of";
+import { fromEvent } from "rxjs/observable/fromEvent";
 
 export interface EthWindow extends Window {
   web3: {
@@ -15,10 +15,10 @@ function getWindow() {
   return window as EthWindow;
 }
 
-export function browserProxyFactory(options?: EthProxyOptions) {
-  const provider$ = Observable.race(
-    Observable.of(getWindow().web3).pipe(filter(x => !!x)),
-    Observable.fromEvent(getWindow(), "load").pipe(map(() => getWindow().web3))
+export function browserProxyFactory(options?: Partial<EthProxyOptions>) {
+  const provider$ = race(
+    of(getWindow().web3).pipe(filter(x => !!x)),
+    fromEvent(getWindow(), "load").pipe(map(() => getWindow().web3))
   ).pipe(filter(x => !!x), map(web3 => web3.currentProvider));
 
   return createProxy(provider$, options);
