@@ -1,5 +1,29 @@
 import * as Web3 from "web3";
 import { Observable } from "rxjs/Observable";
+import { RegisterContractOptions } from "./modules/register-contract";
+
+export type Send<T> = <CK extends keyof T>(
+  nameOrAddress: ContractRef<CK>
+) => <MK extends keyof T[CK]>(method: MK) => T[CK][MK];
+
+export class EthProxy<T = {}> {
+  registerContract: (abi, options: RegisterContractOptions) => void;
+
+  send: Send<T>;
+  query: (queryModel: QueryModel<T>) => Observable<any>;
+  
+  // rxweb3
+  getBalance: (account: string) => Observable<any>;
+  getLatestBlock: () => Observable<Block>;
+  getBlock: (args) => Observable<Block>;
+
+  provider$: Observable<Web3.Provider>;
+  network$: Observable<string>;
+  defaultAccount$: Observable<string | undefined>;
+
+  getContractInfo: (nameOrAddress: string) => Observable<ContractInfo>;
+}
+
 
 export interface InputDefinition {
   indexed: boolean;
@@ -29,9 +53,11 @@ export interface TruffleJson {
 }
 
 export interface TransactionInfo {
+  contractName: string;
   address: string;
-  from: string;
   method: string;
+  txParams: any;
+  args: any;
   tx: string;
 }
 
@@ -52,6 +78,9 @@ export type Transaction =
   | FailedTransaction
   | ConfirmedTransaction;
 
+export type TransactionResult<T> = Observable<Transaction>;
+export type CallResult<T> = Observable<T>;
+
 export interface ContractInfo {
   address: string;
   name: string;
@@ -59,9 +88,20 @@ export interface ContractInfo {
   abi: Web3.AbiDefinition[];
 }
 
+export interface EthProxyInterceptors {
+  call: (obs: Observable<any>) => any;
+  transaction: (obs: Observable<any>) => any;
+  preQuery: (obs: Observable<any>) => any;
+  postQuery: (obs: Observable<any>) => any;
+}
+
 export interface EthProxyOptions {
   pollInterval: number;
   eventReader: (web3: Web3, options: Web3.FilterObject) => Observable<any[]>;
+  store: {
+    dispatch: Function
+  },
+  interceptors: Partial<EthProxyInterceptors>;
 }
 
 export type BlockRange = [number, number];
@@ -77,14 +117,18 @@ export interface QueryResult {
   events: any[];
 }
 
-export interface QueryModel {
+export interface QueryModel<T extends {} = {}> {
   name: string;
   deps: {
-    [contractName: string]: {
-      [eventName: string]: {
-        [inputName: string]: any;
-      } | '*';
-    } | '*';
+    [P in keyof Partial<T>]:
+      | {
+          [eventName: string]:
+            | {
+                [inputName: string]: any;
+              }
+            | "*";
+        }
+      | "*"
   };
 }
 
@@ -93,7 +137,7 @@ export interface Web3Provider {
 }
 
 export interface TransationHashEvent {
-  type: 'tx';
+  type: "tx";
   value: string;
 }
 export interface TransactionConfirmation<T> {
@@ -102,7 +146,7 @@ export interface TransactionConfirmation<T> {
   tx: string;
 }
 export interface TransationConfirmationEvent<T> {
-  type: 'confirmation';
+  type: "confirmation";
   value: TransactionConfirmation<T>;
 }
 export type TransationResultEvent<T> =
@@ -113,7 +157,6 @@ export type ObservableTransactionResult<T> = Observable<
   TransationResultEvent<T>
 >;
 
-
 export interface ContractDefaults {
   from?: string;
   gas?: string;
@@ -121,10 +164,51 @@ export interface ContractDefaults {
   value?: string;
 }
 
-declare module 'rxjs/Observable' {
-  // tslint:disable-next-line:interface-name no-shadowed-variable
-  interface Observable<T> {
-    once(this: Observable<any>, type: "tx", fn: Function): Observable<T>
-    on(this: Observable<any>, type: "confirmation", fn: Function): Observable<T>
-  }
+export type NameRef<T extends string> = T;
+export interface InterfaceRef<T extends string> {
+  interface: T;
+  address: string;
+}
+
+export type ContractRef<T extends string = string> =
+  | NameRef<T>
+  | InterfaceRef<T>;
+
+export interface Block {
+  author: string;
+  difficulty: any;
+  extraData: string;
+  gasLimit: number;
+  gasUsed: number;
+  hash: string;
+  logsBloom: string;
+  miner: string;
+  number: number;
+  parentHash: string;
+  receiptsRoot: string;
+  sealFields: string[];
+  sha3Uncles: string;
+  signature: string;
+  size: number;
+  stateRoot: string;
+  step: string;
+  timestamp: number;
+  totalDifficulty: any;
+  transactions: string[];
+  transactionsRoot: string;
+  uncles: string[];
+}
+
+export interface Provider {
+  sendAsync: any;
+}
+
+export interface EventMetadata {
+  address: string;
+  logIndex: number;
+  transactionHash: string;
+  transactionIndex: number;
+  type: string;
+  blockHash: string;
+  blockNumber: number;
 }
