@@ -2,10 +2,10 @@ export const SET_ACTIVE_ACCOUNT = "SET_ACTIVE_ACCOUNT";
 
 export interface SetActiveAccount {
   type: "SET_ACTIVE_ACCOUNT";
-  payload: string;
+  payload: string | null;
 }
 
-export const createSetActiveAccount = (account: string) => ({
+export const createSetActiveAccount = (account: string | null) => ({
   type: SET_ACTIVE_ACCOUNT,
   payload: account
 });
