@@ -1,6 +1,7 @@
 import { Observable } from "rxjs/Observable";
 import { mergeMapTo, mergeMap, distinctUntilChanged, map } from "rxjs/operators";
 import { ActionsObservable } from "redux-observable";
+import { timer } from "rxjs/observable/timer";
 
 import { getDefaultAccount } from "@eth-proxy/rx-web3";
 import { createSetActiveAccount } from "../actions";
@@ -11,7 +12,7 @@ export const watchAccount = (
   store,
   { web3Proxy$, options }: EpicContext
 ) => {
-  return Observable.timer(0, options.pollInterval).pipe(
+  return timer(0, options.pollInterval).pipe(
     mergeMapTo(web3Proxy$),
     mergeMap(getDefaultAccount),
     distinctUntilChanged(),
