@@ -2,14 +2,14 @@ import * as actions from "../actions";
 import { createSelector } from "reselect";
 
 export type State = {
-  activeAccount?: string;
+  activeAccount?: string | null;
 };
 export function reducer(state: State = {}, action: actions.AccountTypes): State {
   switch (action.type) {
     case actions.SET_ACTIVE_ACCOUNT:
       return {
         ...state,
-        activeAccount: action.payload
+        activeAccount: action.payload || null
       };
     default:
       return state;
